pyproject.toml
[project]
name = "endoscopy-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.2.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
]
